Read the index file ourselves instead of using bgzf_index_load
The problem with this function is that it takes a path and there is
no support or function that take an already fd.
(Which would help immensly as we want to use a fd we get from openat()
of the underlying mountpoint.)

The previous solution was to copy the index files to ~/.fuse-bgzip
but that does not scale.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>